Extra tests of sample matching to disk with mask
